WIP: refactor main into headers and sources
