+++ include/AGIS_Commons.h
@@ -1,0 +1,17 @@
+#ifndef D8C84950_1B28_4594_B461_3D52C4A78A6B
+#define D8C84950_1B28_4594_B461_3D52C4A78A6B
+
+#include <AGIS_Types.h>
+
+extern volatile unsigned int numDrops;
+extern volatile unsigned int dripRate;
+extern volatile float infusedVolume;
+extern volatile unsigned long infusedTime;
+
+extern unsigned int targetVTBI;
+extern unsigned int targetTotalTime;
+extern unsigned int targetDripRate;
+extern unsigned int targetNumDrops;
+extern infusionState_t infusionState;
+
+#endif /* D8C84950_1B28_4594_B461_3D52C4A78A6B */
