+++ include/AGIS_Keypad.h
@@ -4,4 +4,7 @@
 #include <Keypad.h>
 #include <Arduino.h>
 
+#define KEYPAD_ROW_NUM     5
+#define KEYPAD_COLUMN_NUM  4
+
 #endif /* C22DF5A3_1D91_414B_92DD_A70F1FBB8647 */
