+++ include/AGIS_Utilities.h
@@ -1,0 +1,10 @@
+#ifndef A2EE97E5_188E_421B_AEFD_3168005E7645
+#define A2EE97E5_188E_421B_AEFD_3168005E7645
+
+#include <AGIS_Types.h>
+
+const char *getInfusionState(infusionState_t state);
+const char *getMotorState(motorState_t state);
+const char *getButtonState(buttonState_t state);
+
+#endif /* A2EE97E5_188E_421B_AEFD_3168005E7645 */
