--- include/main.h
@@ -1,9 +1,0 @@
-#ifndef D8C84950_1B28_4594_B461_3D52C4A78A6B
-#define D8C84950_1B28_4594_B461_3D52C4A78A6B
-
-volatile unsigned int numDrops = 0;
-volatile unsigned int dripRate = 0;
-volatile float infusedVolume = 0.0f;
-volatile long infusedTime = 0;
-
-#endif /* D8C84950_1B28_4594_B461_3D52C4A78A6B */
